+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -5,8 +5,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.AdminReservationResponse;
-import roomescape.controller.dto.AdminReservationsResponse;
+import roomescape.controller.dto.reservation.AdminReservationResponse;
+import roomescape.controller.dto.reservation.AdminReservationsResponse;
 import roomescape.service.ReservationService;
 
 @RequestMapping("/admin/reservations")
