+++ src/main/java/roomescape/controller/dto/reservation/AdminReservationsResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.reservation;
 
 import java.util.List;
