+++ src/main/java/roomescape/controller/dto/time/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.time;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
